gamesChart folder renamed to lower case, hyphen separated. games-chart
